+++ src/Disks/ObjectStorages/MetadataStorageFromDisk.h
@@ -4,20 +4,10 @@
 
 #include <Disks/IDisk.h>
 #include <Disks/ObjectStorages/DiskObjectStorageMetadata.h>
+#include "MetadataStorageFromDiskTransactionOperations.h"
 
 namespace DB
 {
-
-
-struct IMetadataOperation
-{
-    virtual void execute() = 0;
-    virtual void undo() = 0;
-    virtual void finalize() {}
-    virtual ~IMetadataOperation() = default;
-};
-
-using MetadataOperationPtr = std::unique_ptr<IMetadataOperation>;
 
 enum class MetadataFromDiskTransactionState
 {
